refactor(api): extract route logic to src/lib
Two changes that land in the same files, so they ship together.

Self-fetches removed. Four server components fetched their own public URL
(APP_URL + /api/...), costing an extra edge request and an extra function
invocation per render. Logic moves to src/lib/{explore,profile,repo-stats}-query,
routes become thin wrappers, pages import directly. Same pattern devs/atlas
already used with devs-query. Types stay re-exported from the routes so no
consumer import moves. repo-info is deliberately left alone: it reads x-gh-token,
and its real cost is an uncached contributors fetch, not the round trip.

cacheLife windows widened. All five 60s windows rewrote data that cannot change
that fast: /api/repo-info already pins its GitHub fetch for 300s, profile refresh
carries a 1h cooldown, roadmap voters get fresh tallies in the POST response. The
isPartial branch used the "seconds" profile, whose revalidate floor of 1 is up to
86 400 writes per day per repo, unbounded; it now aligns with the route's own
s-maxage=30.

Two tags were declared and never invalidated, which the 60s windows hid:
repo-stats-* (now fired by badge-update) and profile-* (now fired by the refresh
route). Without those, widening the window would have traded writes for
staleness instead of getting them free.

5 new tests on profile-query, which had none.
